SearchLight.Callbacks.on_exception — Function

Automatically invoked callback when a model exception is triggered internally (ex type conversion exception)

accepts (model <: AbstractModel, ex::TypeConversionException)

returns model <: AbstractModel

SearchLight.Callbacks.on_find — Function

Automatically invoked when model data is retrieved from the DB (on find methods). It will be invoked for each field in the model

accepts (model <: AbstractModel, field_name::Symbol, value::Any)

returns model <: AbstractModel

SearchLight.Callbacks.after_find — Function

Automatically invoked after the object is retrieved from the DB

accepts (model <: AbstractModel)

return model <: AbstractModel

SearchLight.Callbacks.on_save — Function

Automatically invoked when the object is being peristed to the DB

accepts (model <: AbstractModel, field_name::Symbol, value::Any)

returns model <: AbstractModel

SearchLight.Callbacks.before_save — Function

Automatically invoked before the object is being peristed to the DB

accepts (model <: AbstractModel)

returns model <: AbstractModel

SearchLight.Callbacks.after_save — Function

Automatically invoked after the object is being peristed to the DB

accepts (model <: AbstractModel)

returns model <: AbstractModel