SearchLight.Configuration.env — Function

env() :: String

Returns the current environment.

Examples

julia> Configuration.env()
"dev"

SearchLight.Configuration.read_db_connection_data — Function

read_db_connection_data(db_settings_file::String) :: Dict{Any,Any}

Attempts to read the database configuration file and returns the part corresponding to the current environment as a Dict. Does not check if db_settings_file actually exists so it can throw errors. If the database connection information for the current environment does not exist, it returns an empty Dict.

Examples

julia> Configuration.read_db_connection_data(...)
Dict{Any,Any} with 6 entries:
  "host"     => "localhost"
  "password" => "..."
  "username" => "..."
  "port"     => 5432
  "database" => "..."
  "adapter"  => "PostgreSQL"

Missing docstring.

Missing docstring for load. Check Documenter's build log for details.

SearchLight.Configuration.Settings — Type

mutable struct Settings

App configuration - sets up the app's defaults. Individual options are overwritten in the corresponding environment file.