SearchLight - Concise, secure, cross-platform query builder and ORM for Julia

SearchLight.Generator.newmodel — Function

newmodel(name::Union{String,Symbol}; path::Union{String,Nothing} = nothing, pluralize::Bool = true) :: Nothing

Generates a new SearchLight model file and persists it to the resources folder.

SearchLight.Generator.newresource — Function

newresource(resource_name::Union{String,Symbol}) :: Nothing

Generates all the files associated with a new resource and persists them to the resources folder.

SearchLight.Generator.new_table_migration — Function

new_table_migration(migration_name::Union{String,Symbol}; pluralize::Bool = true) :: Nothing

Cretes the migration script for a new table.

SearchLight.Generator.newmigration — Function

newmigration(migration_name::Union{String,Symbol}) :: Nothing

Creates a new migration script.

SearchLight.Generator.setup_resource_path — Function

setup_resource_path(resource_name::String) :: String

Computes and creates the directories structure needed to persist a new resource.

SearchLight.Generator.write_resource_file — Function

write_resource_file(resource_path::String, file_name::String, resource_name::String) :: Bool

Generates all resouce files and persists them to disk.

SearchLight.Generator.newconfig — Function

newconfig(path::String = SearchLight.DB_PATH; filename = SearchLight.SEARCHLIGHT_DB_CONFIG_FILE_NAME) :: Nothing

Creates a new configuration file.

SearchLight.Generator.resource_does_not_exist — Function

resource_does_not_exist(resource_path::String, file_name::String) :: Bool

Checks if a given resource does not exist at resource_path.

SearchLight.Generator.model_file_name — Function

model_file_name(resource_name::Union{String,Symbol}) :: String

Generates the file name for the model corresponding to resource_name.

SearchLight.Generator.validator_file_name — Function

validator_file_name(resource_name::Union{String,Symbol}) :: String

Generates the file name for the validator corresponding to resource_name.