Missing docstring.

Missing docstring for DatabaseMigration. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IrreversibleMigrationException. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ExistingMigrationException. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MigrationNotFoundException. Check Documenter's build log for details.

SearchLight.Migration.new_table — Function

newtable(migration_name::String, resource::String) :: String

Creates a new default migration file and persists it to disk in the configured migrations folder.

SearchLight.Migration.newtable — Function

newtable(migration_name::String, resource::String) :: String

Creates a new default migration file and persists it to disk in the configured migrations folder.

Missing docstring.

Missing docstring for relationship_table_name. Check Documenter's build log for details.

new — Keyword

new

Special function available to inner constructors which created a new object of the type. See the manual section on Inner Constructor Methods for more information.

SearchLight.Migration.migration_hash — Function

migration_hash() :: String

Computes a unique hash for a migration identifier.

SearchLight.Migration.migration_file_name — Function

migration_file_name(migration_name::String) :: String
migration_file_name(cmd_args::Dict{String,Any}, config::Configuration.Settings) :: String

Computes the name of a new migration file.

SearchLight.Migration.migration_module_name — Function

migration_module_name(underscored_migration_name::String) :: String

Computes the name of the module of the migration based on the input from the user (migration name).

SearchLight.Migration.last_up — Function

last_up(; force = false) :: Nothing

Migrates up the last migration. If force is true, the migration will be executed even if it's already up.

SearchLight.Migration.lastup — Function

last_up(; force = false) :: Nothing

Migrates up the last migration. If force is true, the migration will be executed even if it's already up.

SearchLight.Migration.last_down — Function

last_down() :: Nothing

Migrates down the last migration. If force is true, the migration will be executed even if it's already down.

SearchLight.Migration.lastdown — Function

last_down() :: Nothing

Migrates down the last migration. If force is true, the migration will be executed even if it's already down.

SearchLight.Migration.up — Function

up(migration_module_name::String; force::Bool = false) :: Nothing
up_by_module_name(migration_module_name::String; force::Bool = false) :: Nothing

Runs up the migration corresponding to migration_module_name.

Missing docstring.

Missing docstring for up_by_module_name. Check Documenter's build log for details.

SearchLight.Migration.down — Function

down(migration_module_name::String; force::Bool = false) :: Nothing
down_by_module_name(migration_module_name::String; force::Bool = false) :: Nothing

Runs down the migration corresponding to migration_module_name.

Missing docstring.

Missing docstring for down_by_module_name. Check Documenter's build log for details.

SearchLight.Migration.migration_by_module_name — Function

migration_by_module_name(migration_module_name::String) :: Union{Nothing,DatabaseMigration}

Computes the migration that corresponds to migration_module_name.

SearchLight.Migration.all_migrations — Function

all_migrations() :: Tuple{Vector{String},Dict{String,DatabaseMigration}}

Returns the list of all the migrations.

SearchLight.Migration.all — Function

all_migrations() :: Tuple{Vector{String},Dict{String,DatabaseMigration}}

Returns the list of all the migrations.

SearchLight.Migration.last_migration — Function

last_migration() :: DatabaseMigration

Returns the last created migration.

SearchLight.Migration.last — Function

last_migration() :: DatabaseMigration

Returns the last created migration.

SearchLight.Migration.run_migration — Function

run_migration(migration::DatabaseMigration, direction::Symbol; force = false) :: Nothing

Runs migration in up or down, per directon. If force is true, the migration is run regardless of its current status (already up or down).

SearchLight.Migration.store_migration_status — Function

store_migration_status(migration::DatabaseMigration, direction::Symbol) :: Nothing

Persists the direction of the migration into the database.

SearchLight.Migration.upped_migrations — Function

upped_migrations() :: Vector{String}

List of all migrations that are up.

SearchLight.Migration.downed_migrations — Function

downed_migrations() :: Vector{String}

List of all migrations that are down.

SearchLight.Migration.status — Function

status() :: Nothing

Prints a table that displays the direction of each migration.

SearchLight.Migration.all_with_status — Function

all_with_status() :: Tuple{Vector{String},Dict{String,Dict{Symbol,Any}}}

Returns a list of all the migrations and their status.

SearchLight.Migration.all_down!! — Function

all_down!!() :: Nothing

Runs all migrations down.

SearchLight.Migration.all_up!! — Function

all_up!!() :: Nothing

Runs all migrations up.

Missing docstring.

Missing docstring for create_table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for column. Check Documenter's build log for details.

Missing docstring.

Missing docstring for columns. Check Documenter's build log for details.

Missing docstring.

Missing docstring for column_id. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_index. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_indexes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_column. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_columns. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drop_table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_column. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_columns. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_index. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_indexes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for constraint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nextval. Check Documenter's build log for details.

Missing docstring.

Missing docstring for column_id_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drop_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_migrations_table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for init. Check Documenter's build log for details.