SearchLight - Concise, secure, cross-platform query builder and ORM for Julia.

Missing docstring.

Missing docstring for connect. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnect. Check Documenter's build log for details.

Missing docstring.

Missing docstring for connection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UnsupportedException. Check Documenter's build log for details.

DataFrames.DataFrame — Type

DataFrame <: AbstractDataFrame

An AbstractDataFrame that stores a set of named columns

The columns are normally AbstractVectors stored in memory, particularly a Vector or CategoricalVector.

Constructors

DataFrame(pairs::Pair...; makeunique::Bool=false, copycols::Bool=true)
DataFrame(pairs::AbstractVector{<:Pair}; makeunique::Bool=false, copycols::Bool=true)
DataFrame(ds::AbstractDict; copycols::Bool=true)
DataFrame(kwargs..., copycols::Bool=true)

DataFrame(columns::AbstractVecOrMat,
          names::AbstractVector;
          makeunique::Bool=false, copycols::Bool=true)

DataFrame(table; copycols::Union{Bool, Nothing}=nothing)
DataFrame(::DataFrameRow)
DataFrame(::GroupedDataFrame; keepkeys::Bool=true)

Keyword arguments

copycols : whether vectors passed as columns should be copied; by default set to true and the vectors are copied; if set to false then the constructor will still copy the passed columns if it is not possible to construct a DataFrame without materializing new columns. Note the copycols=nothing default in the Tables.jl compatible constructor; it is provided as certain input table types may have already made a copy of columns or the columns may otherwise be immutable, in which case columns are not copied by default. To force a copy in such cases, or to get mutable columns from an immutable input table (like Arrow.Table), pass copycols=true explicitly.

makeunique : if false (the default), an error will be raised

(note that not all constructors support these keyword arguments)

Details on behavior of different constructors

It is allowed to pass a vector of Pairs, a list of Pairs as positional arguments, or a list of keyword arguments. In this case each pair is considered to represent a column name to column value mapping and column name must be a Symbol or string. Alternatively a dictionary can be passed to the constructor in which case its entries are considered to define the column name and column value pairs. If the dictionary is a Dict then column names will be sorted in the returned DataFrame.

In all the constructors described above column value can be a vector which is consumed as is or an object of any other type (except AbstractArray). In the latter case the passed value is automatically repeated to fill a new vector of the appropriate length. As a particular rule values stored in a Ref or a 0-dimensional AbstractArray are unwrapped and treated in the same way.

It is also allowed to pass a vector of vectors or a matrix as as the first argument. In this case the second argument must be a vector of Symbols or strings specifying column names, or the symbol :auto to generate column names x1, x2, ... automatically. Note that in this case if the first argument is a matrix and copycols=false the columns of the created DataFrame will be views of columns the source matrix.

If a single positional argument is passed to a DataFrame constructor then it is assumed to be of type that implements the Tables.jl interface using which the returned DataFrame is materialized.

Finally it is allowed to construct a DataFrame from a DataFrameRow or a GroupedDataFrame. In the latter case the keepkeys keyword argument specifies whether the resulting DataFrame should contain the grouping columns of the passed GroupedDataFrame and the order of rows in the result follows the order of groups in the GroupedDataFrame passed.

Notes

The DataFrame constructor by default copies all columns vectors passed to it. Pass the copycols=false keyword argument (where supported) to reuse vectors without copying them.

By default an error will be raised if duplicates in column names are found. Pass makeunique=true keyword argument (where supported) to accept duplicate names, in which case they will be suffixed with _i (i starting at 1 for the first duplicate).

If an AbstractRange is passed to a DataFrame constructor as a column it is always collected to a Vector (even if copycols=false). As a general rule AbstractRange values are always materialized to a Vector by all functions in DataFrames.jl before being stored in a DataFrame.

DataFrame can store only columns that use 1-based indexing. Attempting to store a vector using non-standard indexing raises an error.

The DataFrame type is designed to allow column types to vary and to be dynamically changed also after it is constructed. Therefore DataFrames are not type stable. For performance-critical code that requires type-stability either use the functionality provided by select/transform/combine functions, use Tables.columntable and Tables.namedtupleiterator functions, use barrier functions, or provide type assertions to the variables that hold columns extracted from a DataFrame.

Examples

julia> DataFrame((a=[1, 2], b=[3, 4])) # Tables.jl table constructor
2×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      3
   2 │     2      4

julia> DataFrame([(a=1, b=0), (a=2, b=0)]) # Tables.jl table constructor
2×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      0
   2 │     2      0

julia> DataFrame("a" => 1:2, "b" => 0) # Pair constructor
2×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      0
   2 │     2      0

julia> DataFrame([:a => 1:2, :b => 0]) # vector of Pairs constructor
2×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      0
   2 │     2      0

julia> DataFrame(Dict(:a => 1:2, :b => 0)) # dictionary constructor
2×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      0
   2 │     2      0

julia> DataFrame(a=1:2, b=0) # keyword argument constructor
2×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      0
   2 │     2      0

julia> DataFrame([[1, 2], [0, 0]], [:a, :b]) # vector of vectors constructor
2×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      0
   2 │     2      0

julia> DataFrame([1 0; 2 0], :auto) # matrix constructor
2×2 DataFrame
 Row │ x1     x2
     │ Int64  Int64
─────┼──────────────
   1 │     1      0
   2 │     2      0

Missing docstring.

Missing docstring for find. Check Documenter's build log for details.

Missing docstring.

Missing docstring for onereduce. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findone. Check Documenter's build log for details.

Base.one — Function

one(x)
one(T::type)

Return a multiplicative identity for x: a value such that one(x)*x == x*one(x) == x. Alternatively one(T) can take a type T, in which case one returns a multiplicative identity for any x of type T.

If possible, one(x) returns a value of the same type as x, and one(T) returns a value of type T. However, this may not be the case for types representing dimensionful quantities (e.g. time in days), since the multiplicative identity must be dimensionless. In that case, one(x) should return an identity value of the same precision (and shape, for matrices) as x.

If you want a quantity that is of the same type as x, or of type T, even if x is dimensionful, use oneunit instead.

See also the identity function, and I in LinearAlgebra for the identity matrix.

Examples

julia> one(3.7)
1.0

julia> one(Int)
1

julia> import Dates; one(Dates.Day(1))
1

Missing docstring.

Missing docstring for randone. Check Documenter's build log for details.

Base.all — Function

all(itr) -> Bool

Test whether all elements of a boolean collection are true, returning false as soon as the first false value in itr is encountered (short-circuiting). To short-circuit on true, use any.

If the input contains missing values, return missing if all non-missing values are true (or equivalently, if the input contains no false value), following three-valued logic.

See also: all!, any, count, &, , &&, allunique.

Examples

julia> a = [true,false,false,true]
4-element Vector{Bool}:
 1
 0
 0
 1

julia> all(a)
false

julia> all((println(i); v) for (i, v) in enumerate(a))
1
2
false

julia> all([missing, false])
false

julia> all([true, missing])
missing

all(p, itr) -> Bool

Determine whether predicate p returns true for all elements of itr, returning false as soon as the first item in itr for which p returns false is encountered (short-circuiting). To short-circuit on true, use any.

If the input contains missing values, return missing if all non-missing values are true (or equivalently, if the input contains no false value), following three-valued logic.

Examples

julia> all(i->(4<=i<=6), [4,5,6])
true

julia> all(i -> (println(i); i < 3), 1:10)
1
2
3
false

julia> all(i -> i > 0, [1, missing])
missing

julia> all(i -> i > 0, [-1, missing])
false

julia> all(i -> i > 0, [1, 2])
true

all(A; dims)

Test whether all values along the given dimensions of an array are true.

Examples

julia> A = [true false; true true]
2×2 Matrix{Bool}:
 1  0
 1  1

julia> all(A, dims=1)
1×2 Matrix{Bool}:
 1  0

julia> all(A, dims=2)
2×1 Matrix{Bool}:
 0
 1

all(p, A; dims)

Determine whether predicate p returns true for all elements along the given dimensions of an array.

Examples

julia> A = [1 -1; 2 2]
2×2 Matrix{Int64}:
 1  -1
 2   2

julia> all(i -> i > 0, A, dims=1)
1×2 Matrix{Bool}:
 1  0

julia> all(i -> i > 0, A, dims=2)
2×1 Matrix{Bool}:
 0
 1

Base.first — Function

first(coll)

Get the first element of an iterable collection. Return the start point of an AbstractRange even if it is empty.

See also: only, firstindex, last.

Examples

julia> first(2:2:10)
2

julia> first([1; 2; 3; 4])
1

first(itr, n::Integer)

Get the first n elements of the iterable collection itr, or fewer elements if itr is not long enough.

See also: startswith, Iterators.take.

Julia 1.6

This method requires at least Julia 1.6.

Examples

julia> first(["foo", "bar", "qux"], 2)
2-element Vector{String}:
 "foo"
 "bar"

julia> first(1:6, 10)
1:6

julia> first(Bool[], 1)
Bool[]

first(s::AbstractString, n::Integer)

Get a string consisting of the first n characters of s.

Examples

julia> first("∀ϵ≠0: ϵ²>0", 0)
""

julia> first("∀ϵ≠0: ϵ²>0", 1)
"∀"

julia> first("∀ϵ≠0: ϵ²>0", 3)
"∀ϵ≠"

first(q::Deque)

Returns the first element of the deque q.

first(D::CircularDeque)

Get the item at the front of the queue.

first(s::Stack)

Get the top item from the stack. Sometimes called peek.

first(h::BinaryHeap)

Returns the element at the top of the heap h.

first(h::BinaryMinMaxHeap)

Get the first (minimum) of the heap.

first(sc)

Argument sc is a SortedDict, SortedMultiDict or SortedSet. This function returns the first item (a k=>v pair for SortedDict and SortedMultiDict or a key for SortedSet) according to the sorted order in the container. Thus, first(sc) is equivalent to deref((sc,startof(sc))). It is an error to call this function on an empty container. Time: O(log n)

first(sc)

Argument sc is a SortedDict, SortedMultiDict or SortedSet. This function returns the first item (a k=>v pair for SortedDict and SortedMultiDict or a key for SortedSet) according to the sorted order in the container. Thus, first(sc) is equivalent to deref((sc,startof(sc))). It is an error to call this function on an empty container. Time: O(log n)

first(sc)

Argument sc is a SortedDict, SortedMultiDict or SortedSet. This function returns the first item (a k=>v pair for SortedDict and SortedMultiDict or a key for SortedSet) according to the sorted order in the container. Thus, first(sc) is equivalent to deref((sc,startof(sc))). It is an error to call this function on an empty container. Time: O(log n)

first(cb::CircularBuffer)

Get the first element of CircularBuffer.

first(df::AbstractDataFrame)

Get the first row of df as a DataFrameRow.

first(df::AbstractDataFrame, n::Integer; view::Bool=false)

Get a data frame with the n first rows of df.

If view=false a freshly allocated DataFrame is returned. If view=true then a SubDataFrame view into df is returned.

Base.last — Function

last(coll)

Get the last element of an ordered collection, if it can be computed in O(1) time. This is accomplished by calling lastindex to get the last index. Return the end point of an AbstractRange even if it is empty.

See also first, endswith.

Examples

julia> last(1:2:10)
9

julia> last([1; 2; 3; 4])
4

last(itr, n::Integer)

Get the last n elements of the iterable collection itr, or fewer elements if itr is not long enough.

Julia 1.6

This method requires at least Julia 1.6.

Examples

julia> last(["foo", "bar", "qux"], 2)
2-element Vector{String}:
 "bar"
 "qux"

julia> last(1:6, 10)
1:6

julia> last(Float64[], 1)
Float64[]

last(s::AbstractString, n::Integer)

Get a string consisting of the last n characters of s.

Examples

julia> last("∀ϵ≠0: ϵ²>0", 0)
""

julia> last("∀ϵ≠0: ϵ²>0", 1)
"0"

julia> last("∀ϵ≠0: ϵ²>0", 3)
"²>0"

last(q::Deque)

Returns the last element of the deque q.

last(D::CircularDeque)

Get the item from the back of the queue.

last(sc)

Argument sc is a SortedDict, SortedMultiDict or SortedSet. This function returns the last item (a k=>v pair for SortedDict and SortedMultiDict or a key for SortedSet) according to the sorted order in the container. Thus, last(sc) is equivalent to deref((sc,lastindex(sc))). It is an error to call this function on an empty container. Time: O(log n)

last(sc)

Argument sc is a SortedDict, SortedMultiDict or SortedSet. This function returns the last item (a k=>v pair for SortedDict and SortedMultiDict or a key for SortedSet) according to the sorted order in the container. Thus, last(sc) is equivalent to deref((sc,lastindex(sc))). It is an error to call this function on an empty container. Time: O(log n)

last(sc)

Argument sc is a SortedDict, SortedMultiDict or SortedSet. This function returns the last item (a k=>v pair for SortedDict and SortedMultiDict or a key for SortedSet) according to the sorted order in the container. Thus, last(sc) is equivalent to deref((sc,lastindex(sc))). It is an error to call this function on an empty container. Time: O(log n)

last(cb::CircularBuffer)

Get the last element of CircularBuffer.

last(df::AbstractDataFrame)

Get the last row of df as a DataFrameRow.

last(df::AbstractDataFrame, n::Integer; view::Bool=false)

Get a data frame with the n last rows of df.

If view=false a freshly allocated DataFrame is returned. If view=true then a SubDataFrame view into df is returned.

Missing docstring.

Missing docstring for save. Check Documenter's build log for details.

Missing docstring.

Missing docstring for save!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for save!!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for _save!!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for updatewith!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for convertmethod. Check Documenter's build log for details.

Missing docstring.

Missing docstring for booltypes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for autoconvert. Check Documenter's build log for details.

Missing docstring.

Missing docstring for updatewith!!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for updatewith. Check Documenter's build log for details.

Missing docstring.

Missing docstring for createwith. Check Documenter's build log for details.

Missing docstring.

Missing docstring for updateby_or_create. Check Documenter's build log for details.

Missing docstring.

Missing docstring for update_or_create. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findone_or_create. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_models. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_model. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_model!!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_select_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_from_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_where_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_order_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_group_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_limit_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_offset_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_having_part. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_join_part. Check Documenter's build log for details.

SearchLight.columns_from_joins — Function

columnsfromjoins(joins::Vector{SQLJoin})::Vector{SQLColumn}

Extracts columns from joins param and adds to be used for the SELECT part

Missing docstring.

Missing docstring for column_data_to_column_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for prepare_column_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for columns_names_by_table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframes_by_table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_find_sql. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_fetch_sql. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_store_sql. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_sqlinput. Check Documenter's build log for details.

Missing docstring.

Missing docstring for delete_all. Check Documenter's build log for details.

Missing docstring.

Missing docstring for deleteall. Check Documenter's build log for details.

Missing docstring.

Missing docstring for delete. Check Documenter's build log for details.

Missing docstring.

Missing docstring for query. Check Documenter's build log for details.

Missing docstring.

Missing docstring for clone. Check Documenter's build log for details.

Missing docstring.

Missing docstring for columns. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ispersisted. Check Documenter's build log for details.

Missing docstring.

Missing docstring for column_field_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for persistable_fields. Check Documenter's build log for details.

Missing docstring.

Missing docstring for settable_fields. Check Documenter's build log for details.

Missing docstring.

Missing docstring for table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pk. Check Documenter's build log for details.

Missing docstring.

Missing docstring for primary_key_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for strip_table_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for is_fully_qualified. Check Documenter's build log for details.

Missing docstring.

Missing docstring for from_fully_qualified. Check Documenter's build log for details.

Missing docstring.

Missing docstring for strip_module_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_fully_qualified. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_sql_column_names. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_sql_column_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_fully_qualified_sql_column_names. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fo_fully_qualified_sql_column_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for from_literal_column_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_dict. Check Documenter's build log for details.

Missing docstring.

Missing docstring for to_string_dict. Check Documenter's build log for details.

Missing docstring.

Missing docstring for enclosure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for update_query_part. Check Documenter's build log for details.

SearchLight.escape_column_name — Function

escape_column_name(c::SQLColumn) :: SQLColumn
escape_column_name(s::String)

Sanitizes input to be use as column names in SQL queries.

SearchLight.escape_value — Function

escape_value(i::SQLInput)

Sanitizes input to be used as values in SQL queries.

Missing docstring.

Missing docstring for index_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sql. Check Documenter's build log for details.

SearchLight.add_quotes — Function

add_quotes(str::String) :: String

Adds quotes around str and escapes any previously existing quotes.

SearchLight.strip_quotes — Function

strip_quotes(str::String) :: String

Unquotes str.

SearchLight.isquoted — Function

isquoted(str::String) :: Bool

Checks weather or not str is quoted.

SearchLight.expand_nullable — Function

expand_nullable{T}(value::Union{Nothing,T}, default::T) :: T

Returns value if it is not nothing - otherwise default.

Core.NamedTuple — Type

NamedTuple

NamedTuples are, as their name suggests, named Tuples. That is, they're a tuple-like collection of values, where each entry has a unique name, represented as a Symbol. Like Tuples, NamedTuples are immutable; neither the names nor the values can be modified in place after construction.

Accessing the value associated with a name in a named tuple can be done using field access syntax, e.g. x.a, or using getindex, e.g. x[:a] or x[(:a, :b)]. A tuple of the names can be obtained using keys, and a tuple of the values can be obtained using values.

Note

Iteration over NamedTuples produces the values without the names. (See example below.) To iterate over the name-value pairs, use the pairs function.

The @NamedTuple macro can be used for conveniently declaring NamedTuple types.

Examples

julia> x = (a=1, b=2)
(a = 1, b = 2)

julia> x.a
1

julia> x[:a]
1

julia> x[(:a,)]
(a = 1,)

julia> keys(x)
(:a, :b)

julia> values(x)
(1, 2)

julia> collect(x)
2-element Vector{Int64}:
 1
 2

julia> collect(pairs(x))
2-element Vector{Pair{Symbol, Int64}}:
 :a => 1
 :b => 2

In a similar fashion as to how one can define keyword arguments programmatically, a named tuple can be created by giving a pair name::Symbol => value or splatting an iterator yielding such pairs after a semicolon inside a tuple literal:

julia> (; :a => 1)
(a = 1,)

julia> keys = (:a, :b, :c); values = (1, 2, 3);

julia> (; zip(keys, values)...)
(a = 1, b = 2, c = 3)

As in keyword arguments, identifiers and dot expressions imply names:

julia> x = 0
0

julia> t = (; x)
(x = 0,)

julia> (; t.x)
(x = 0,)

Julia 1.5

Implicit names from identifiers and dot expressions are available as of Julia 1.5.

Julia 1.7

Use of getindex methods with multiple Symbols is available as of Julia 1.7.